Refactored handlers, deleted jar.rs
